fix: move payment block to card component
