+++ src/components/form/FormDonation/PaymentForm/PaymentForm.tsx
@@ -1,16 +1,12 @@
 import { useState } from 'react';
-import { useTranslation } from 'react-i18next';
 import {
   FieldValues,
   FormProvider,
   SubmitHandler,
   useForm,
 } from 'react-hook-form';
-import { useMediaQuery } from 'react-responsive';
-import { ButtonApp, Modal } from '@components/UI';
+import { Modal } from '@components/UI';
 import PaymentBlock from '../PaymentBlock/PaymentBlock';
-import Email from '../../Email/Email';
-import Comment from '../../Comment/Comment';
 import ChooseProject from './ChooseProject';
 import Payment from '@components/payment/Payment/Payment';
 import s from './PaymentForm.module.scss';
@@ -34,16 +30,10 @@ const initialDataForm = {
 
 const PaymentForm = ({ content }: PaymentFormProps) => {
   const [dataForm, setDataForm] = useState<IDataForm>(initialDataForm);
-  const { t } = useTranslation();
   const [isOpenModal, setIsOpenModal] = useState(false);
   const methods = useForm({
     mode: 'onChange',
   });
-  const isMobile = useMediaQuery({ minWidth: 480, maxWidth: 767 });
-
-  const {
-    formState: { isValid },
-  } = methods;
 
   const onSubmit = (data: any) => {
     console.log(data);
@@ -69,16 +59,6 @@ const PaymentForm = ({ content }: PaymentFormProps) => {
       >
         {content === 'donateProject' && <ChooseProject />}
         <PaymentBlock />
-        <Email className={s.form__email} />
-        <Comment title='Comment' placeholder='Type here...' />
-        <ButtonApp
-          type='submit'
-          size={isMobile ? 'Xlarge' : 'medium'}
-          className={s.form__button}
-          disabled={!isValid}
-        >
-          {t('Pay urgent!')}
-        </ButtonApp>
       </form>
     </FormProvider>
   );
